requirements.txt
asgiref==3.9.1
contourpy==1.3.3
cycler==0.12.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
fonttools==4.59.2
kiwisolver==1.4.9
matplotlib==3.10.6
narwhals==2.4.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.16.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
google-generativeai==0.8.0
